Backend/ParkingSpaceFinder/requirements.txt
asgiref==3.7.2
auth-token==0.0.3
Django==5.0.3
django-cors-headers==4.3.1
django-rest-authtoken==2.1.4
djangorestframework==3.15.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
